requirements.txt
streamlit>=1.10.0
gspread
google-auth
pandas
openpyxl